refactor: reduce Code Text package surface
