+++ packages/sdk-components-registry/package.json
@@ -43,9 +43,11 @@
     "@webstudio-is/sdk": "workspace:*",
     "@webstudio-is/sdk-components-animation": "workspace:*",
     "@webstudio-is/sdk-components-react": "workspace:*",
-    "@webstudio-is/sdk-components-react-radix": "workspace:*"
+    "@webstudio-is/sdk-components-react-radix": "workspace:*",
+    "shiki": "4.4.1"
   },
   "devDependencies": {
+    "@testing-library/react": "^16.2.0",
     "@types/react": "^18.2.70",
     "@types/react-dom": "^18.2.25",
     "@webstudio-is/tsconfig": "workspace:*",
