+++ packages/sdk-components-registry/src/code-text-canvas.test.tsx
